share routes
