+++ packages/server/index.d.ts
@@ -14,7 +14,6 @@
 
 // types for the `server` package
 export namespace CyServer {
-  export type getRemoteState = () => RemoteState
 
   // TODO: pull this from main types
   export interface Config {
@@ -26,18 +25,6 @@ export namespace CyServer {
      * URL to Cypress's runner.
      */
     responseTimeout: number
-  }
-
-  export interface RemoteState {
-    auth?: {
-      username: string
-      password: string
-    }
-    domainName: string
-    strategy: 'file' | 'http'
-    origin: string
-    fileServer: string
-    visiting: string
   }
 
   export interface Socket {
